home page data moved to separate file
